Update Odroid N2 boot and genimage with 2020.07 uboot

+++ board/batocera/post-image-script.sh
@@ -173,7 +173,7 @@ case "${BATOCERA_TARGET}" in
 
 	S912)
 	MKIMAGE=${HOST_DIR}/bin/mkimage
-	MKBOOTIMAGE=${HOST_DIR}/bin/mkbootimg
+	#MKBOOTIMAGE=${HOST_DIR}/bin/mkbootimg
 	BOARD_DIR="${BATO_DIR}/amlogic/s912"
 	# boot
 	rm -rf "${BINARIES_DIR:?}/boot"      || exit 1
@@ -183,9 +183,14 @@ case "${BATOCERA_TARGET}" in
 	$MKIMAGE -C none -A arm64 -T script -d "${BOARD_DIR}/boot/aml_autoscript.txt" "${BINARIES_DIR}/boot/aml_autoscript"
 	cp "${BOARD_DIR}/boot/aml_autoscript.zip" "${BINARIES_DIR}/boot"     || exit 1
 	cp "${BINARIES_DIR}/batocera-boot.conf" "${BINARIES_DIR}/boot/batocera-boot.conf" || exit 1
-	cp "${BINARIES_DIR}/all_merged.dtb" "${BINARIES_DIR}/dtb.img" || exit 1
-	$MKBOOTIMAGE --kernel "${BINARIES_DIR}/Image" --ramdisk "${BINARIES_DIR}/initrd" --second "${BINARIES_DIR}/dtb.img" --output "${BINARIES_DIR}/linux" || exit 1
-	cp "${BINARIES_DIR}/Image" "${BINARIES_DIR}/boot/boot/linux" || exit 1
+	#cp "${BINARIES_DIR}/all_merged.dtb" "${BINARIES_DIR}/dtb.img" || exit 1
+	for DTB in meson-gxm-khadas-vim2 meson-gxm-nexbox-a1 meson-gxm-q200 meson-gxm-q201 meson-gxm-rbox-pro meson-gxm-vega-s96 meson-gxm-s912-libretech-pc
+        do
+                cp "${BINARIES_DIR}/${DTB}.dtb" "${BINARIES_DIR}/boot/boot" || exit 1
+        done
+        "${HOST_DIR}/bin/mkimage" -A arm64 -O linux -T kernel -C none -a 0x1080000 -e 0x1080000 -n 5.x -d "${BINARIES_DIR}/Image" "${BINARIES_DIR}/uImage" || exit 1
+	#$MKBOOTIMAGE --kernel "${BINARIES_DIR}/Image" --ramdisk "${BINARIES_DIR}/initrd" --second "${BINARIES_DIR}/meson-gxm-khadas-vim2.dtb" --output "${BINARIES_DIR}/linux" || exit 1
+	cp "${BINARIES_DIR}/uImage" "${BINARIES_DIR}/boot/boot/linux" || exit 1
 
 	cp "${BINARIES_DIR}/rootfs.squashfs" "${BINARIES_DIR}/boot/boot/batocera.update" || exit 1
 	cp -pr "${BINARIES_DIR}/tools"       "${BINARIES_DIR}/boot/"                || exit 1
@@ -332,39 +337,45 @@ case "${BATOCERA_TARGET}" in
 	;;
 
 	ODROIDN2)
-	BOARD_DIR="${BATO_DIR}/amlogic/odroidn2"
-	# /boot
-	rm -rf "${BINARIES_DIR}/boot"            || exit 1
-	mkdir -p "${BINARIES_DIR}/boot/boot"     || exit 1
-	"${HOST_DIR}/bin/mkimage" -A arm64 -O linux -T kernel -C none -a 0x1080000 -e 0x1080000 -n 5.x -d "${BINARIES_DIR}/Image" "${BINARIES_DIR}/uImage" || exit 1
-	cp "${BINARIES_DIR}/uImage"                            "${BINARIES_DIR}/boot/boot/linux"                            || exit 1
-	cp "${BINARIES_DIR}/u-boot.bin"                        "${BINARIES_DIR}/boot/u-boot.bin"                            || exit 1
-	cp "${BINARIES_DIR}/uInitrd"                           "${BINARIES_DIR}/boot/boot/uInitrd"                          || exit 1
-	cp "${BINARIES_DIR}/rootfs.squashfs"                   "${BINARIES_DIR}/boot/boot/batocera.update"                  || exit 1
+        BOARD_DIR="${BATO_DIR}/amlogic/odroidn2"
+        MKIMAGE=${HOST_DIR}/bin/mkimage
+
+        # boot
+        rm -rf "${BINARIES_DIR:?}/boot"      || exit 1
+        mkdir -p "${BINARIES_DIR}/boot/boot" || exit 1
+        mkdir -p "${BINARIES_DIR}/boot/extlinux" || exit 1
+        "${HOST_DIR}/bin/mkimage" -A arm64 -O linux -T kernel -C none -a 0x1080000 -e 0x1080000 -n linux -d "${BINARIES_DIR}/Image" "${BINARIES_DIR}/uImage" || exit 1
+        cp "${BOARD_DIR}/boot/boot-logo.bmp.gz"                "${BINARIES_DIR}/boot"                                       || exit 1
+        cp "${BINARIES_DIR}/batocera-boot.conf"                "${BINARIES_DIR}/boot/batocera-boot.conf"                    || exit 1
+        cp "${BINARIES_DIR}/uImage"                            "${BINARIES_DIR}/boot/boot/linux"                            || exit 1
 	cp "${BINARIES_DIR}/meson-g12b-odroid-n2.dtb"          "${BINARIES_DIR}/boot/boot/meson-g12b-odroid-n2.dtb"         || exit 1
 	cp "${BINARIES_DIR}/meson-g12b-odroid-n2-plus.dtb"     "${BINARIES_DIR}/boot/boot/meson-g12b-odroid-n2_plus.dtb"    || exit 1
-	cp "${BINARIES_DIR}/batocera-boot.conf"                "${BINARIES_DIR}/boot/batocera-boot.conf"                    || exit 1
+        cp "${BINARIES_DIR}/initrd.gz"                         "${BINARIES_DIR}/boot/boot/initrd.gz"                        || exit 1
+        cp "${BINARIES_DIR}/rootfs.squashfs"                   "${BINARIES_DIR}/boot/boot/batocera.update"                  || exit 1
+        cp "${BOARD_DIR}/boot/extlinux.conf"                   "${BINARIES_DIR}/boot/extlinux"                              || exit 1
+	cp "${BINARIES_DIR}/u-boot.bin"                        "${BINARIES_DIR}/boot/u-boot.bin"                            || exit 1
 	cp "${BOARD_DIR}/boot/extlinux.conf"                   "${BINARIES_DIR}/boot/boot/extlinux.conf"                    || exit 1
 	cp "${BOARD_DIR}/boot/boot.ini"                        "${BINARIES_DIR}/boot/boot.ini"                              || exit 1
 	cp "${BOARD_DIR}/boot/config.ini"                      "${BINARIES_DIR}/boot/config.ini"                            || exit 1
-	cp "${BOARD_DIR}/boot/boot-logo.bmp.gz"                "${BINARIES_DIR}/boot"                                       || exit 1
-	cp -pr "${BINARIES_DIR}/tools"                         "${BINARIES_DIR}/boot/" || exit 1
+        cp -pr "${BINARIES_DIR}/tools"                         "${BINARIES_DIR}/boot/" || exit 1
+        cp "${BINARIES_DIR}/u-boot.bin.sd.bin"                  "${BINARIES_DIR}/boot/" || exit 1
 
-	# boot.tar.xz
-	echo "creating boot.tar.xz"
-	(cd "${BINARIES_DIR}/boot" && tar -I "xz -T0" -cf "${BATOCERA_BINARIES_DIR}/boot.tar.xz" tools boot batocera-boot.conf boot-logo.bmp.gz boot.ini config.ini) || exit 1
+        # boot.tar.xz
+        echo "creating boot.tar.xz"
+        (cd "${BINARIES_DIR}/boot" && tar -I "xz -T0" -cf "${BATOCERA_BINARIES_DIR}/boot.tar.xz" extlinux tools boot batocera-boot.conf boot-logo.bmp.gz boot.ini config.ini) || exit 1
 
-	# batocera.img
-	mv "${BINARIES_DIR}/boot/boot/batocera.update" "${BINARIES_DIR}/boot/boot/batocera" || exit 1
-	GENIMAGE_TMP="${BUILD_DIR}/genimage.tmp"
-	BATOCERAIMG="${BATOCERA_BINARIES_DIR}/batocera.img"
-	rm -rf "${GENIMAGE_TMP}" || exit 1
-	cp "${BOARD_DIR}/genimage.cfg" "${BINARIES_DIR}" || exit 1
-	echo "generating image"
-	genimage --rootpath="${TARGET_DIR}" --inputpath="${BINARIES_DIR}/boot" --outputpath="${BATOCERA_BINARIES_DIR}" --config="${BINARIES_DIR}/genimage.cfg" --tmppath="${GENIMAGE_TMP}" || exit 1
-	rm -f "${BATOCERA_BINARIES_DIR}/boot.vfat" || exit 1
-	rm -f "${BATOCERA_BINARIES_DIR}/userdata.ext4" || exit 1
-	sync || exit 1
+        # batocera.img
+        # rename the squashfs : the .update is the version that will be renamed at boot to replace the old version
+        mv "${BINARIES_DIR}/boot/boot/batocera.update" "${BINARIES_DIR}/boot/boot/batocera" || exit 1
+        GENIMAGE_TMP="${BUILD_DIR}/genimage.tmp"
+        BATOCERAIMG="${BATOCERA_BINARIES_DIR}/batocera.img"
+        rm -rf "${GENIMAGE_TMP}" || exit 1
+        cp "${BOARD_DIR}/genimage.cfg" "${BINARIES_DIR}" || exit 1
+        echo "generating image"
+        genimage --rootpath="${TARGET_DIR}" --inputpath="${BINARIES_DIR}/boot" --outputpath="${BATOCERA_BINARIES_DIR}" --config="${BINARIES_DIR}/genimage.cfg" --tmppath="${GENIMAGE_TMP}" || exit 1
+        rm -f "${BATOCERA_BINARIES_DIR}/boot.vfat" || exit 1
+        rm -f "${BATOCERA_BINARIES_DIR}/userdata.ext4" || exit 1
+        sync || exit 1
 	;;
 
 	ODROIDGOA)
